Create JamStackPlayer.js
The Jam Stack Player is html 5 audio shuffle player that works great for our Jam Stacks. When a user adds a song to their Stack or Jam Stack. We used to stack tapes or CDs and when one would finish we would load another one. The Jam Stacker does that whole shuffle play think for us online.