Update Header.js
Changed Logo Head Graphic to Take Users To Login If Not Logged in and if logged in and clicked to swipe which the main page of the website. Also if user is logged in their profile image is displayed top right. I need to add a login log out feature but I think I will do that from the simple user profile.